Revert "t6026-merge-attr: don't fail if sleep exits early"
This reverts commit c96eb0a7ba95f3baea6464e09ca39c4f2a0d3b3a.

The point of the test is that the stray process was still running
when 'git merge' did its thing through its completion, so a failure
to "kill" it means we didn't give a condition to the test to trigger
a possible future breakage.  Appending "|| :" to the "kill" is
sweeping a test-bug under the rug.